backend/requirements/base.txt
Django==5.0.9
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-axes==6.4.0
psycopg2-binary==2.9.9
redis==5.0.8
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.6.0
cryptography==42.0.8
stripe==9.12.0
python-decouple==3.8
Pillow==10.4.0
gunicorn==22.0.0
